refactored code, corrected logic in url controller. Deleted redis/newLinkDto

+++ src/urls/urls.controller.ts
@@ -17,7 +17,6 @@ import { UserUrlService } from './urls.service';
 
 @Controller('urls')
 export class UrlsController {
-
   @Inject(UserUrlService)
   private userUrlService: UserUrlService;
 
@@ -29,11 +28,9 @@ export class UrlsController {
 
   @Post()
   public async createUrl(@Body() body: urlDto): Promise<UserUrl> {
-
     try {
-
       if (await this.redisService.shortUrlExist(body.short_link)) {
-        throw "Short URL unavailable :(";
+        throw 'Short URL unavailable :(';
       }
 
       const userId = this.requestService.getUser().id;
@@ -46,19 +43,18 @@ export class UrlsController {
       if (isAvailable) {
         return this.userUrlService.addUrlToUser(userId, body);
       } else {
-        throw "Short URL unavailable :(";
+        throw 'Short URL unavailable :(';
       }
-
     } catch (err: any) {
       throw new HttpException(err, 409);
     }
-
   }
 
   //////////////////////////////////// TODO: ver si queda
   @Get('/:short_link/clicks')
   public async getLinkClicks(
-    @Param('short_link') short_link: string): Promise<number> {
+    @Param('short_link') short_link: string,
+  ): Promise<number> {
     return await this.redisService.getUrlClicks(short_link);
   }
   ////////////////////////////////////
@@ -68,30 +64,25 @@ export class UrlsController {
     @Param('short_link') short_link: string,
     @Body() body: updateUrlDto,
   ): Promise<UserUrl> {
-
     try {
-
       const userId = this.requestService.getUser().id;
       const url = await this.getUserUrlById(short_link);
 
       if (url) {
-
-        let updatedRedis = await this.redisService.updateUrlKeyValue(short_link, body.long_link);
+        let updatedRedis = await this.redisService.updateUrlKeyValue(
+          short_link,
+          body.long_link,
+        );
 
         if (updatedRedis) {
           return this.userUrlService.updateUrl(userId, body, url);
         } else {
-          throw "An was encountered trying to update the link, please try again later";
+          throw 'An was encountered trying to update the link, please try again later';
         }
-
       }
-
     } catch (err: any) {
-
       throw new HttpException(err, 409);
-
     }
-
   }
   @Get()
   public async getUserUrls(): Promise<UserUrl> {
@@ -107,14 +98,22 @@ export class UrlsController {
       const url = userUrls.urls.find(
         (element) => element.short_link == short_link,
       );
-      return url;
+      if (url) {
+        return url;
+      }
+      throw new HttpException('Short link not found :(', 404);
     }
   }
 
   @Delete('/:short_link')
   public async deleteUrlById(@Param('short_link') short_link: string) {
     const userId = this.requestService.getUser().id;
-    this.userUrlService.deleteUrlFromUser(userId, short_link);
-    this.redisService.deleteKeyValue(short_link);
+    const deleted = await this.userUrlService.deleteUrlFromUser(
+      userId,
+      short_link,
+    );
+    if (deleted) {
+      this.redisService.deleteKeyValue(short_link);
+    }
   }
 }
